Adapt to API changes in ad-hoc stereotype registration
The registration now requires a source object to be handed into the definition creation to keep track of the definition's origin. JsonNodeHandler is now made aware of a StereotypeCatalog to be able to obtain the definition of a stereotype. That makes the source available in handleStereotype(…) which now needs actual handling of it.